Create a Socket.io server that uses the same HTTP server instance as Nest.js
